test(ci): cut CPU test tail — drop dataset_num_proc to 1, split builder tests
The Python 3.12 PyTest legs run ~2x slower than 3.14 on the same test set
(816s vs 403s) and were tipping over the 30-minute job timeout. Two causes,
both in the slow tail:

- dataset_num_proc=4 forks 4 dataset workers per .map() on CPU-only runners,
  each re-importing the torch stack to process a few hundred rows — pure
  overhead. Lower to 1 in the affected tests (none assert on it or test
  multiprocessing); results are unchanged.
- --dist loadfile pins a whole file to one worker, so the entire builder
  suite serialized on a single worker at the end. Move shared fixtures to
  tests/core/conftest.py and split the RL trainer-builder tests into
  test_builders_rl.py so they run on a separate worker from the SFT/reward
  builder tests.
