+++ tests/test_packed_dataset.py
@@ -55,7 +55,7 @@ class TestPacking(unittest.TestCase):
                         "type": "alpaca",
                     },
                 ],
-                "dataset_num_proc": 4,
+                "dataset_num_proc": 1,
                 "num_epochs": 1,
                 "max_steps": 20,
                 "save_steps": 10,
